+++ src/module/gimbal/mod_gimbal.hpp
@@ -66,7 +66,7 @@ class Gimbal {
 
   void SetMode(Mode mode);
 
-  uint32_t last_wakeup_;
+  float last_wakeup_;
 
   float now_;
